# Better memory layout, more vectorization, Numba