I made a few additional experiments:

- I stored the polygon vertices sequentially in memory, which allows for full vectorization with NumPy. This made things faster, but still slower than Rust.
- I tested JIT compilation with [Numba](https://numba.pydata.org/). This was faster faster than the Rust implementation.
- I again tested Numba, but this time I discarded polygons early if the center was so far away that no vertex can possibly be within range of the query point. This was much faster.

I'd also like to mention the [KDTree data structure](https://docs.scipy.org/doc/scipy/reference/generated/scipy.spatial.KDTree.html) which supports ball point queries. This only pays of for many points, so the current $O(n^2)$ approach is still faster, but in general, this would be the way to go.

### My results

```
Measuring `poly_match_v1`
Took an avg of 223.03ms per iteration (14 iterations)
Measuring `poly_match_v1_5_vectorized`
Took an avg of 41.36ms per iteration (73 iterations)
Measuring `poly_match_v1_6_excessively_vectorized`
Took an avg of 10.04ms per iteration (299 iterations)
Measuring `poly_match_v1_7_numba`
Took an avg of 1.95ms per iteration (500 iterations)
Measuring `poly_match_v1_8_numba_early_exit`
Took an avg of 0.40ms per iteration (500 iterations)
Measuring `poly_match_v2_riir`
Took an avg of 562.50ms per iteration (6 iterations)
Measuring `poly_match_v3_riir_polygon`
Took an avg of 164.46ms per iteration (19 iterations)
Measuring `poly_match_v4_avoid_allocs`
Took an avg of 47.12ms per iteration (64 iterations)
Measuring `poly_match_v5_1_varied_polygons`
Took an avg of 3.19ms per iteration (500 iterations)
Measuring `poly_match_v5_find_all_in_rust`
Took an avg of 3.11ms per iteration (500 iterations)
```

For my own experiments, I tested polygons with different shape and varying number of vertices. I believe that this makes the test more realistic.

I could also update your experiments with the new polygons. It might change the performance numbers slightly.

![polygons](https://github.com/ohadravid/poly-match/assets/18725165/0b452a59-45c3-4c8a-a215-42d872c24de4)