@ohadravid 2023-06-24
Hi @99991 !

Those are very cool examples, the numba one esp.

I can't say that I completely follow everything that's going on, but: like I wrote in the original article, we aren't really doing "useful" work in this library, it is just a proxy to the original library's perf problems. 

Because of that, some of the optimizations don't really make sense: For example, a KD Tree won't work for the simple reason that both polygons and points change each time (again, just to maintain the analogy).

Similarly, doing extra computation per polygon before the actual search is "cheating" a bit, because (a) we have to include it when measuring the perf and (b) because it wasn't really possible in the original library (to clarify, analogous algorithmic improvements were used for the first stages of the 3D pipeline, but we still had the "list of objects with numpy attributes" problem when dealing with the results)

To me, I'd love to merge one or both of:

1. An example which is "as close as possible" to the original, but which uses numba with little algorithmic changes.
2. An example which "takes numpy to the limit", but which is clearly measuring the same thing (starting with just the list of polygons).

Of course, you can also keep this open PR as-is. It is interesting in it's own right :)