@99991 2023-06-24
> Those are very cool examples, the numba one esp.

Thank you!

> Because of that, some of the optimizations don't really make sense: For example, a KD Tree won't work for the simple reason that both polygons and points change each time (again, just to maintain the analogy).

KDTrees can be built relatively quickly. It might still be worth it. Or perhaps not. I do not know the original task.

> we still had the "list of objects with numpy attributes" problem when dealing with the results

Copying the polygon data into a single NumPy array so that it can be worked on efficiently takes about 2 milliseconds. This means that a benchmark which includes this time is mostly measuring the overhead of an inefficient data layout. That is still a valid problem which might come up in practical applications, but I am not very interested in solving it.