@99991 2023-06-24
I noticed that I am computing the distance to the closest vertex of the polygon instead of to the center, making my code about 5 times slower than it needed to be. I should probably have started with writing unit tests.