@ohadravid 2023-06-24
> This means that a benchmark which includes this time is mostly measuring the overhead of an inefficient data layout.

The article is essentially about converting an inefficient data layout in Python to a more efficient one in Rust, while keeping the same API 😉 

Part of the constraints of the system is that the previous steps produce this layout (mostly due to being very efficient at some other task), which means that we have to include any conversion costs when comparing results.

As I said, this is still a very interesting solution 🚀 , it's just that we can't directly compare the results.