@99991 2023-06-24
> Part of the constraints of the system is that the previous steps produce this layout (mostly due to being very efficient at some other task)

I am fairly certain that this previous step could also be modified to use a contiguous memory layout, but it might be too much work.