@ohadravid 2023-06-30
BTW, I noticed that your `poly_match_v2_riir` is slower than the original Python version.
Perhphse you didn't use the `--release` flag when building using `maturin`?

My results on your branch are:
```
$ python ./measure_all.py
Measuring `poly_match_v1`
Took an avg of 295.71ms per iteration (11 iterations)
Measuring `poly_match_v1_5_vectorized`
Took an avg of 44.52ms per iteration (68 iterations)
Measuring `poly_match_v1_6_excessively_vectorized`
Took an avg of 9.17ms per iteration (328 iterations)
Measuring `poly_match_v1_7_numba`
Took an avg of 2.19ms per iteration (500 iterations)
Measuring `poly_match_v1_8_numba_early_exit`
Took an avg of 0.51ms per iteration (500 iterations)
Measuring `poly_match_v2_riir`
Took an avg of 26.49ms per iteration (114 iterations)
Measuring `poly_match_v3_riir_polygon`
Took an avg of 6.91ms per iteration (435 iterations)
Measuring `poly_match_v4_avoid_allocs`
Took an avg of 3.43ms per iteration (500 iterations)
Measuring `poly_match_v5_1_varied_polygons`
Took an avg of 1.05ms per iteration (500 iterations)
Measuring `poly_match_v5_find_all_in_rust`
Took an avg of 1.22ms per iteration (500 iterations)
```