@99991 2023-06-30
You are right, I did not run it with the release flag. Perhaps build instructions would be helpful.

Here are the results with `--release`.

```
Measuring `poly_match_v1`
Took an avg of 225.51ms per iteration (14 iterations)
Measuring `poly_match_v1_5_vectorized`
Took an avg of 34.61ms per iteration (87 iterations)
Measuring `poly_match_v1_6_excessively_vectorized`
Took an avg of 7.63ms per iteration (394 iterations)
Measuring `poly_match_v1_7_numba`
Took an avg of 1.50ms per iteration (500 iterations)
Measuring `poly_match_v1_8_numba_early_exit`
Took an avg of 0.31ms per iteration (500 iterations)
Measuring `poly_match_v2_riir`
Took an avg of 20.56ms per iteration (146 iterations)
Measuring `poly_match_v3_riir_polygon`
Took an avg of 5.58ms per iteration (500 iterations)
Measuring `poly_match_v4_avoid_allocs`
Took an avg of 2.88ms per iteration (500 iterations)
Measuring `poly_match_v5_1_varied_polygons`
Took an avg of 0.89ms per iteration (500 iterations)
Measuring `poly_match_v5_find_all_in_rust`
Took an avg of 0.88ms per iteration (500 iterations)
```